// lib/model/batchRetrieveByKeysReq.ts
/**
 * 配置服务API
 * 配置服务API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from '../api';

export class BatchRetrieveByKeysReq {
    /**
    * keys
    */
    'keys'?: Array<string>;
    /**
    * 资源ID
    */
    'resourceId'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "keys",
            "baseName": "keys",
            "type": "Array<string>"
        },
        {
            "name": "resourceId",
            "baseName": "resourceId",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return BatchRetrieveByKeysReq.attributeTypeMap;
    }
}