// lib/model/batchRetrieveByResourcesReq.ts
/**
 * 配置服务API
 * 配置服务API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from '../api';

export class BatchRetrieveByResourcesReq {
    /**
    * key
    */
    'key'?: string;
    /**
    * 资源ID组
    */
    'resourceIds'?: Array<string>;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "key",
            "baseName": "key",
            "type": "string"
        },
        {
            "name": "resourceIds",
            "baseName": "resourceIds",
            "type": "Array<string>"
        }    ];

    static getAttributeTypeMap() {
        return BatchRetrieveByResourcesReq.attributeTypeMap;
    }
}