// lib/model/batchRetrieveConfigTemplateByKeysReq.ts
/**
 * 配置服务API
 * 配置服务API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from '../api';

export class BatchRetrieveConfigTemplateByKeysReq {
    'keys'?: Array<string>;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "keys",
            "baseName": "keys",
            "type": "Array<string>"
        }    ];

    static getAttributeTypeMap() {
        return BatchRetrieveConfigTemplateByKeysReq.attributeTypeMap;
    }
}