// lib/model/cursorExtra.ts
/**
 * 配置服务API
 * 配置服务API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from '../api';

export class CursorExtra {
    /**
    * 大小
    */
    'size'?: number;
    /**
    * 查询方向，0表示游标前，1表示游标后
    */
    'direction'?: number;
    /**
    * 资源ID
    */
    'resourceId'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "size",
            "baseName": "size",
            "type": "number"
        },
        {
            "name": "direction",
            "baseName": "direction",
            "type": "number"
        },
        {
            "name": "resourceId",
            "baseName": "resourceId",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return CursorExtra.attributeTypeMap;
    }
}